[Chapter13] Fix a bug in the previous patch.
The bug causes the fleet drop immediately aftering reach
the right edege. Becauses we didn't change the fleet direction
properly. This is because if the direction is not changed in time,
all the aliens will reach the right edge and each adds a
fleet_drop_speed to the alien.rect.y, this quickly adds up, causing
the fleet dropping down very quickly.
Also reset the dropping speed to 10 to make is less boring, otherwise
it is dropping so slow.